+++ .idea/compiler.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="CompilerConfiguration">
+    <annotationProcessing>
+      <profile name="Maven default annotation processors profile" enabled="true">
+        <sourceOutputDir name="target/generated-sources/annotations" />
+        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
+        <outputRelativeToContentRoot value="true" />
+        <module name="arandela-calculator" />
+        <module name="acosta-calculator" />
+        <module name="api-demo (1) (com.badosa)" />
+        <module name="batangan-calculator" />
+        <module name="api-demo (1) (com.rodenas.springboot)" />
+        <module name="rodenas-calculator" />
+        <module name="calculator-alonde" />
+        <module name="isles-calculator" />
+        <module name="galindon-calculator" />
+        <module name="api-demo" />
+        <module name="calculator-dizon" />
+        <module name="alvarez-calculator" />
+        <module name="bautista-calculator" />
+        <module name="calculator-nepomuceno" />
+        <module name="api-demo (1) (com.costiniano)" />
+      </profile>
+    </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="api.demo" target="21" />
+    </bytecodeTargetLevel>
+  </component>
+  <component name="JavacSettings">
+    <option name="ADDITIONAL_OPTIONS_OVERRIDE">
+      <module name="api-demo" options="-parameters" />
+      <module name="api-demo (1) (com.badosa)" options="-parameters" />
+      <module name="api-demo (1) (com.costiniano)" options="-parameters" />
+      <module name="api-demo (1) (com.rodenas.springboot)" options="-parameters" />
+      <module name="api.demo" options="-parameters" />
+    </option>
+  </component>
+</project>
