--- springboot-lab/DevTeam1/arandela/api-demo/src/main/java/org/springframework/boot/SpringApplication.java
@@ -1,8 +1,0 @@
-package org.springframework.boot;
-
-public class SpringApplication {
-    public static void run(Class<?> primarySource, String[] args) {
-        // Mock implementation for javac compilation
-    }
-}
-
