--- springboot-lab/DevTeam1/arandela/api-demo/src/main/java/org/springframework/boot/autoconfigure/SpringBootApplication.java
@@ -1,12 +1,0 @@
-package org.springframework.boot.autoconfigure;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SpringBootApplication {
-}
-
