--- springboot-lab/DevTeam1/arandela/api-demo/src/main/java/org/springframework/web/bind/annotation/GetMapping.java
@@ -1,13 +1,0 @@
-package org.springframework.web.bind.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface GetMapping {
-    String[] value() default {};
-}
-
