+++ springboot-lab/DevTeam1/arandela/api-demo/src/main/resources/application.properties
@@ -9,5 +9,3 @@ app.version=1.0.0
 
 # Logging
 logging.level.com.arandela.springboot=DEBUG
-
-
