--- springboot-lab/DevTeam1/arandela/api-demo/src/test/java/org/junit/jupiter/api/Test.java
@@ -1,12 +1,0 @@
-package org.junit.jupiter.api;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Test {
-}
-
