+++ springboot-lab/DevTeam1/nepomuceno/api-demo/src/main/java/com/nepomuceno/springboot/ApiDemoApplication.java
@@ -1,4 +1,4 @@
-package com.arandela.springboot;
+package com.nepomuceno.springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
