+++ springboot-lab/DevTeam1/nepomuceno/api-demo/src/main/resources/application.properties
@@ -1,11 +1,10 @@
 spring.application.name=api-demo
-
 # Server configuration
 server.port=8080
 
 # Application info
-app.name=Arandela Spring Boot API
+app.name=Nepomuceno Spring Boot API
 app.version=1.0.0
 
 # Logging
-logging.level.com.arandela.springboot=DEBUG
+logging.level.com.nepomuceno.springboot=DEBUG
