+++ springboot-lab/DevTeam1/nepomuceno/api-demo/src/test/java/com/nepomuceno/springboot/ApiDemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.arandela.springboot;
+package com.nepomuceno.springboot;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
