+++ springboot-lab/DevTeam1/nepomuceno/api-demo/src/test/java/com/nepomuceno/springboot/MeControlTest.java
@@ -1,0 +1,29 @@
+package com.nepomuceno.springboot;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureWebMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.setup.MockMvcBuilders.*;
+
+@SpringBootTest
+@AutoConfigureWebMvc
+public class MeControllerTest {
+
+    @Autowired
+    private MeController meController;
+
+    private MockMvc mockMvc = standaloneSetup(meController).build();
+
+    @Test
+    public void testGetMe() throws Exception {
+        mockMvc.perform(MockMvcRequestBuilders.get("/me"))
+                .andExpect(status().isOk())
+                .andExpect(content().contentType("application/json"))
+                .andExpect(jsonPath("$.name").exists());
+    }
+}
